Added packege crema.entropy. Moved AbellanEntropy and MaximumEntropy in this package.
